pyproject.toml
[project]
name = "laser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pyserial>=3.5",
]
